Allow marfs_write() to handle O_CREAT, even though fuse shouldn't use that.
Also changed the return type from int to ssize_t, because that's what it was actually
returning (given that the <size> argument implies we might be called with a buffer
having a number of elements that requires a size_t to specify).

+++ fuse/src/marfs_ops.h
@@ -175,8 +175,8 @@ int  marfs_utimensat(const char* path, const struct timespec times[2], int flags
 
 int  marfs_utimens(const char* path, const struct timespec ts[2]);
 
-int  marfs_write(const char* path, const char* buf, size_t size, off_t offset,
-                 MarFS_FileHandle* fh);
+ssize_t  marfs_write(const char* path, const char* buf, size_t size, off_t offset,
+                     MarFS_FileHandle* fh);
